# src/Dle.Control/openapi/dle-control.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Deep Link Engine — control plane",
    "description": "Link, domain, application, tenant and credential management, plus attribution and analytics reporting. Errors are RFC 9457 problem documents whose type member is one of the stable identifiers documented at https://docs.dle.dev/problems/. Writes accept an Idempotency-Key header: the same key with the same body replays the stored response, and the same key with a different body is a conflict.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/analytics/clicks": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Handles the time series report.",
        "description": "The points carry clicks, installs and conversions together because that is what the port\r\nreturns: the three come out of one pass over the same window, and splitting them into three\r\nrequests would triple the work to produce numbers that have to line up anyway.",
        "operationId": "GetClickAnalytics",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/installs": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Handles the time series report.",
        "description": "The points carry clicks, installs and conversions together because that is what the port\r\nreturns: the three come out of one pass over the same window, and splitting them into three\r\nrequests would triple the work to produce numbers that have to line up anyway.",
        "operationId": "GetInstallAnalytics",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/breakdown": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "One dimension of the window, grouped and ranked.",
        "description": "The dimension is resolved against a fixed allowlist before any query is built, so an unknown name is a validation error rather than a failed query.",
        "operationId": "GetAnalyticsBreakdown",
        "parameters": [
          {
            "name": "dimension",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakdownResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/funnels": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Handles the funnel report: clicks to installs to attributed installs to conversions.",
        "description": "The totals for the window, and the conversion rate they imply.",
        "operationId": "GetAnalyticsFunnel",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnelSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/attribution-quality": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Handles the attribution quality report — the deterministic, probabilistic and unmatched\r\nsplit with the mean confidence of the guesses (ADR-008, §0.2).",
        "description": "How many installs were matched with certainty, how many were guessed and at what mean confidence, and how many were not matched at all. The unmatched share is published rather than hidden: a measurement product that only shows what it caught cannot be checked, and this number is the argument against the commercial measurement partners rather than a footnote (ADR-008, §0.2).",
        "operationId": "GetAttributionQuality",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributionQualityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/export": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "A report as a CSV or Parquet download.",
        "description": "Streams the report rather than buffering it, so a two year breakdown is a download and not an outage (FR-203).",
        "operationId": "ExportAnalytics",
        "parameters": [
          {
            "name": "report",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dimension",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/stream": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "A live view of the window over Server-Sent Events.",
        "description": "Emits a snapshot whenever the totals move and a heartbeat when they do not. The click stream is written by the edge process, so this polls the analytics store rather than forwarding an in-process event; the shape is already a stream of snapshots, so replacing the poll with a subscription later changes nothing a client can see (FR-206).",
        "operationId": "StreamAnalytics",
        "parameters": [
          {
            "name": "intervalSeconds",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/tenant": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Handles an export.",
        "description": "GDPR article 20 portability and the DORA article 30 exit plan are the same feature, and this is it: newline delimited JSON per table in a ZIP, streamed straight out of the database. Credential material — key hashes, webhook secrets, account hashes — is never included, and the manifest says so explicitly rather than leaving the gap to be discovered (FR-249, §E.6.4).",
        "operationId": "ExportTenantData",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttributionQualityResponse": {
        "required": [
          "deterministic",
          "probabilistic",
          "unmatched",
          "average_probabilistic_confidence"
        ],
        "type": "object",
        "properties": {
          "deterministic": {
            "type": "integer",
            "description": "Attributed installs whose match is deterministic, confidence exactly 1.00.",
            "format": "int64"
          },
          "probabilistic": {
            "type": "integer",
            "description": "Attributed installs matched probabilistically.",
            "format": "int64"
          },
          "unmatched": {
            "type": "integer",
            "description": "Installs that could not be attributed at all.",
            "format": "int64"
          },
          "average_probabilistic_confidence": {
            "type": "number",
            "description": "Mean confidence across the probabilistic matches only.",
            "format": "double"
          },
          "by_match_type": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchTypeSummary"
            },
            "description": "Per strategy detail."
          }
        },
        "description": "The honest attribution split. This is the number commercial measurement partners systematically\r\nblur, and stating it plainly is the product argument, not a footnote (ADR-008, §0.2)."
      },
      "BreakdownResponse": {
        "required": [
          "dimension"
        ],
        "type": "object",
        "properties": {
          "dimension": {
            "type": "string",
            "description": "Dimension that was grouped by, for example `country` or `channel`."
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreakdownRow"
            },
            "description": "The rows, most significant first."
          }
        },
        "description": "A breakdown of one dimension over the queried window."
      },
      "BreakdownRow": {
        "required": [
          "key",
          "clicks",
          "installs",
          "value"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Value of the grouped dimension. Never `null`: rows whose\r\n            dimension is unknown are reported under the literal key `unknown` rather than dropped,\r\n            so the sum of the rows always matches the totals."
          },
          "clicks": {
            "type": "integer",
            "description": "Number of clicks in the group.",
            "format": "int64"
          },
          "installs": {
            "type": "integer",
            "description": "Number of installs attributed to the group.",
            "format": "int64"
          },
          "value": {
            "type": [
              "null",
              "number"
            ],
            "description": "Summed monetary value of conversions in the group, or\r\n            `null` when the group produced no conversion carrying a value.",
            "format": "double"
          }
        },
        "description": "One row of a dimensional breakdown report (FR-202), for example clicks per country or per\r\ncampaign. Rows are ordered by long BreakdownRow.Clicks descending and truncated to\r\nint AnalyticsQuery.Limit."
      },
...
      "FunnelSummary": {
        "required": [
          "clicks",
          "installs",
          "attributed",
          "conversions",
          "conversion_rate"
        ],
        "type": "object",
        "properties": {
          "clicks": {
            "type": "integer",
            "description": "Human clicks in the reported interval.",
            "format": "int64"
          },
          "installs": {
            "type": "integer",
            "description": "First application opens in the reported interval.",
            "format": "int64"
          },
          "attributed": {
            "type": "integer",
            "description": "Installs that were matched to a click by any strategy. Always less\r\n            than or equal to Installs.",
            "format": "int64"
          },
          "conversions": {
            "type": "integer",
            "description": "Conversion events reported by the SDK in the interval.",
            "format": "int64"
          },
          "conversion_rate": {
            "type": "number",
            "description": "Conversions divided by clicks, rounded to four decimal places.\r\n            Zero when there were no clicks; the value is never extrapolated.",
            "format": "double"
          }
        },
        "description": "Aggregated click to conversion funnel for one AnalyticsQuery (FR-202)."
      },
...
      "MatchTypeSummary": {
        "required": [
          "match_type",
          "count",
          "average_confidence"
        ],
        "type": "object",
        "properties": {
          "match_type": {
            "type": "string",
            "description": "Strategy name, one of the constants on\r\n            `Dle.Domain.Attribution.MatchTypeNames`."
          },
          "count": {
            "type": "integer",
            "description": "Number of attribution records produced by that strategy.",
            "format": "int64"
          },
          "average_confidence": {
            "type": "number",
            "description": "Mean confidence of those records, rounded to two decimal\r\n            places. Exactly `1.00` for every deterministic strategy.",
            "format": "double"
          }
        },
        "description": "Distribution of attribution strategies over the reported interval — the data behind the\r\nhonest dashboard demanded by ADR-008 (\"78 % deterministic, 14 % probabilistic with an average\r\nconfidence of 0.71, 8 % unmatched\")."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TimeSeriesPoint": {
        "required": [
          "bucket",
          "clicks",
          "installs",
          "conversions"
        ],
        "type": "object",
        "properties": {
          "bucket": {
            "type": "string",
            "description": "Inclusive start of the bucket in UTC.",
            "format": "date-time"
          },
          "clicks": {
            "type": "integer",
            "description": "Number of human clicks in the bucket. Crawler traffic is excluded unless\r\n            bool AnalyticsQuery.IncludeBots is set (FR-205).",
            "format": "int64"
          },
          "installs": {
            "type": "integer",
            "description": "Number of first application opens attributed to the bucket.",
            "format": "int64"
          },
          "conversions": {
            "type": "integer",
            "description": "Number of conversion events reported by the SDK in the bucket.",
            "format": "int64"
          }
        },
        "description": "One bucket of a time series report (FR-202). Buckets are aligned in UTC according to\r\nTimeGrain AnalyticsQuery.Grain and are returned in ascending order."
      },
      "TimeSeriesResponse": {
        "required": [
          "grain"
        ],
        "type": "object",
        "properties": {
          "grain": {
            "type": "string",
            "description": "Bucket size actually applied."
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeSeriesPoint"
            },
            "description": "The buckets, oldest first."
          },
          "totals": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Totals across the window.",
                "$ref": "#/components/schemas/FunnelSummary"
              }
            ]
          }
        },
        "description": "A time series plus the totals for the same window."
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "description": "A control-plane API key, presented as a bearer token or in the X-Api-Key header. The key looks like dle_<prefix>_<secret>; only its Argon2id hash is stored, so a lost key is replaced rather than recovered.",
        "scheme": "bearer"
      },
      "SdkKey": {
        "type": "apiKey",
        "description": "A key embedded in a customer application. It authenticates the two SDK ingestion routes and nothing else, and must be assumed extractable from the application binary.",
        "name": "X-Dle-Sdk-Key",
        "in": "header"
      }
    }
  },
...
}